WriteValue now uses a UInt16 in the parameter

+++ source/System.Devices.Dac/IDacChannel.cs
@@ -1,5 +1,5 @@
 ﻿//
-// Copyright (c) 2017 The nanoFramework project contributors
+// Copyright (c) 2019 The nanoFramework project contributors
 // See LICENSE file in the project root for full license information.
 //
 
@@ -7,7 +7,7 @@ namespace System.Devices.Dac
 {
     internal interface IDacChannel
     {
-        bool WriteValue(double value);
+        bool WriteValue(UInt16 value);
 
         DacController Controller { get; }
     }
